+++ src/Kindly/Functor.hs
@@ -69,7 +69,12 @@ import Data.Monoid qualified as Monoid
 import Data.Ord (Down)
 import Data.Profunctor qualified as Hask.Profunctor
 import Data.Profunctor.Cayley qualified as Hask.Profunctor
+import Data.Profunctor.Choice qualified as Hask.Profunctor
+import Data.Profunctor.Closed qualified as Hask.Profunctor
 import Data.Profunctor.Composition qualified as Hask.Profunctor
+import Data.Profunctor.Mapping qualified as Hask.Profunctor
+import Data.Profunctor.Strong qualified as Hask.Profunctor
+import Data.Profunctor.Traversing qualified as Hask.Profunctor
 import Data.Profunctor.Yoneda qualified as Hask.Profunctor
 import Data.Proxy (Proxy)
 import Data.Semigroup qualified as Semigroup
@@ -564,6 +569,91 @@ instance (FunctorOf (->) (->) f, FunctorOf (->) (->) (p a)) => CategoricalFuncto
   type Cod (Hask.Profunctor.Cayley f p a) = (->)
 
   map g (Hask.Profunctor.Cayley fp) = Hask.Profunctor.Cayley $ map (map g) fp
+
+instance (forall x. FunctorOf (->) (->) (p x)) => CategoricalFunctor (Hask.Profunctor.Tambara p a) where
+  type Dom (Hask.Profunctor.Tambara p a) = (->)
+  type Cod (Hask.Profunctor.Tambara p a) = (->)
+
+  map f (Hask.Profunctor.Tambara t) = Hask.Profunctor.Tambara $ map (\(b, c) -> (f b, c)) t
+
+instance CategoricalFunctor (Hask.Profunctor.Pastro p a) where
+  type Dom (Hask.Profunctor.Pastro p a) = (->)
+  type Cod (Hask.Profunctor.Pastro p a) = (->)
+
+  map f (Hask.Profunctor.Pastro l m r) = Hask.Profunctor.Pastro (f . l) m r
+
+instance CategoricalFunctor (Hask.Profunctor.Cotambara q a) where
+  type Dom (Hask.Profunctor.Cotambara q a) = (->)
+  type Cod (Hask.Profunctor.Cotambara q a) = (->)
+
+  map f (Hask.Profunctor.Cotambara n r) = Hask.Profunctor.Cotambara n (Hask.Profunctor.rmap f r)
+
+instance CategoricalFunctor (Hask.Profunctor.Copastro p a) where
+  type Dom (Hask.Profunctor.Copastro p a) = (->)
+  type Cod (Hask.Profunctor.Copastro p a) = (->)
+
+  map f (Hask.Profunctor.Copastro g) = Hask.Profunctor.Copastro $ \n -> Hask.Profunctor.rmap f (g n)
+
+instance (forall x. FunctorOf (->) (->) (p x)) => CategoricalFunctor (Hask.Profunctor.TambaraSum p a) where
+  type Dom (Hask.Profunctor.TambaraSum p a) = (->)
+  type Cod (Hask.Profunctor.TambaraSum p a) = (->)
+
+  map f (Hask.Profunctor.TambaraSum t) =
+    Hask.Profunctor.TambaraSum $ map (\e -> case e of Left b -> Left (f b); Right c -> Right c) t
+
+instance CategoricalFunctor (Hask.Profunctor.PastroSum p a) where
+  type Dom (Hask.Profunctor.PastroSum p a) = (->)
+  type Cod (Hask.Profunctor.PastroSum p a) = (->)
+
+  map f (Hask.Profunctor.PastroSum l m r) = Hask.Profunctor.PastroSum (f . l) m r
+
+instance CategoricalFunctor (Hask.Profunctor.CotambaraSum q a) where
+  type Dom (Hask.Profunctor.CotambaraSum q a) = (->)
+  type Cod (Hask.Profunctor.CotambaraSum q a) = (->)
+
+  map f (Hask.Profunctor.CotambaraSum n r) = Hask.Profunctor.CotambaraSum n (Hask.Profunctor.rmap f r)
+
+instance CategoricalFunctor (Hask.Profunctor.CopastroSum p a) where
+  type Dom (Hask.Profunctor.CopastroSum p a) = (->)
+  type Cod (Hask.Profunctor.CopastroSum p a) = (->)
+
+  map f (Hask.Profunctor.CopastroSum g) = Hask.Profunctor.CopastroSum $ \n -> Hask.Profunctor.rmap f (g n)
+
+instance (forall x. FunctorOf (->) (->) (p x)) => CategoricalFunctor (Hask.Profunctor.Closure p a) where
+  type Dom (Hask.Profunctor.Closure p a) = (->)
+  type Cod (Hask.Profunctor.Closure p a) = (->)
+
+  map f (Hask.Profunctor.Closure t) = Hask.Profunctor.Closure $ map (f .) t
+
+instance CategoricalFunctor (Hask.Profunctor.Environment p a) where
+  type Dom (Hask.Profunctor.Environment p a) = (->)
+  type Cod (Hask.Profunctor.Environment p a) = (->)
+
+  map f (Hask.Profunctor.Environment l m r) = Hask.Profunctor.Environment (f . l) m r
+
+instance CategoricalFunctor (Hask.Profunctor.FreeTraversing p a) where
+  type Dom (Hask.Profunctor.FreeTraversing p a) = (->)
+  type Cod (Hask.Profunctor.FreeTraversing p a) = (->)
+
+  map f (Hask.Profunctor.FreeTraversing l m r) = Hask.Profunctor.FreeTraversing (f . l) m r
+
+instance (forall x. FunctorOf (->) (->) (p x)) => CategoricalFunctor (Hask.Profunctor.CofreeTraversing p a) where
+  type Dom (Hask.Profunctor.CofreeTraversing p a) = (->)
+  type Cod (Hask.Profunctor.CofreeTraversing p a) = (->)
+
+  map f (Hask.Profunctor.CofreeTraversing t) = Hask.Profunctor.CofreeTraversing $ map (Hask.fmap f) t
+
+instance CategoricalFunctor (Hask.Profunctor.FreeMapping p a) where
+  type Dom (Hask.Profunctor.FreeMapping p a) = (->)
+  type Cod (Hask.Profunctor.FreeMapping p a) = (->)
+
+  map f (Hask.Profunctor.FreeMapping l m r) = Hask.Profunctor.FreeMapping (f . l) m r
+
+instance (forall x. FunctorOf (->) (->) (p x)) => CategoricalFunctor (Hask.Profunctor.CofreeMapping p a) where
+  type Dom (Hask.Profunctor.CofreeMapping p a) = (->)
+  type Cod (Hask.Profunctor.CofreeMapping p a) = (->)
+
+  map f (Hask.Profunctor.CofreeMapping t) = Hask.Profunctor.CofreeMapping $ map (Hask.fmap f) t
 
 --------------------------------------------------------------------------------